+++ src/algorithms/PVT/gnuradio_blocks/rtklib_pvt_gs.cc
@@ -1254,8 +1254,9 @@ void rtklib_pvt_gs::msg_handler_telemetry(const pmt::pmt_t& msg)
                         }
                     if (gps_cnav_ephemeris->signal_health != 0)
                         {
-                            std::cout << TEXT_RED << "Satellite " << Gnss_Satellite(std::string("GPS"), gps_cnav_ephemeris->PRN)
-                                      << " is not healthy, not used for navigation" << TEXT_RESET << '\n';
+                            std::cout << "Satellite " << Gnss_Satellite(std::string("GPS"), gps_cnav_ephemeris->PRN)
+                                      << " does not report a healthy status in the CNAV message,"
+                                      << " use PVT solutions at your own risk.\n";
                         }
                     DLOG(INFO) << "New GPS CNAV ephemeris record has arrived ";
                 }
@@ -2023,7 +2024,7 @@ int rtklib_pvt_gs::work(int noutput_items, gr_vector_const_void_star& input_item
                             if (tmp_eph_iter_cnav != d_internal_pvt_solver->gps_cnav_ephemeris_map.cend())
                                 {
                                     const uint32_t prn_aux = tmp_eph_iter_cnav->second.PRN;
-                                    if ((prn_aux == in[i][epoch].PRN) && (((std::string(in[i][epoch].Signal) == std::string("2S")) || (std::string(in[i][epoch].Signal) == std::string("L5"))) && (tmp_eph_iter_cnav->second.signal_health == 0)))
+                                    if ((prn_aux == in[i][epoch].PRN) && (((std::string(in[i][epoch].Signal) == std::string("2S")) || (std::string(in[i][epoch].Signal) == std::string("L5")))))
                                         {
                                             store_valid_observable = true;
                                         }
@@ -2114,10 +2115,10 @@ int rtklib_pvt_gs::work(int noutput_items, gr_vector_const_void_star& input_item
                         {
                             const double Rx_clock_offset_s = d_internal_pvt_solver->get_time_offset_s();
 
-                            //**************** time tags ****************
+                            // **************** time tags ****************
                             if (d_enable_rx_clock_correction == false)  // todo: currently only works if clock correction is disabled (computed clock offset is applied here)
                                 {
-                                    //************ Source TimeTag comparison with GNSS computed TOW *************
+                                    // ************ Source TimeTag comparison with GNSS computed TOW *************
 
                                     if (!d_TimeChannelTagTimestamps.empty())
                                         {
@@ -2151,7 +2152,7 @@ int rtklib_pvt_gs::work(int noutput_items, gr_vector_const_void_star& input_item
                                                 }
                                         }
                                 }
-                            //**********************************************
+                            // **********************************************
 
                             if (fabs(Rx_clock_offset_s) * 1000.0 > d_max_obs_block_rx_clock_offset_ms)
                                 {