+++ src/algorithms/signal_source/libs/rtl_tcp_dongle_info.cc
@@ -33,11 +33,11 @@ boost::system::error_code Rtl_Tcp_Dongle_Info::read(boost::asio::ip::tcp::socket
 {
     boost::system::error_code ec;
 
-    unsigned char data[sizeof(char) * 4 + sizeof(uint32_t) * 2];
+    std::vector<unsigned char> data(sizeof(char) * 4 + sizeof(uint32_t) * 2);
     size_t received_bits = socket.receive(boost::asio::buffer(data), 0, ec);
     if (!ec && (received_bits > 0))
         {
-            std::memcpy(magic_, data, 4);
+            std::memcpy(magic_, data.data(), 4);
 
             uint32_t type;
             std::memcpy(&type, &data[4], 4);