+++ docs/CHANGELOG.md
@@ -12,7 +12,38 @@ SPDX-FileCopyrightText: 2011-2022 Carles Fernandez-Prades <carles.fernandez@cttc
 
 All notable changes to GNSS-SDR will be documented in this file.
 
+## [GNSS-SDR v0.0.17](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.17) - 2022-04-20
+
+### Improvements in Availability:
+
+- Compute PVT solutions when using GPS L5 signals even if the satellite is
+  reported as not healthy in the CNAV message.
+
+### Improvements in Portability:
+
+- Updated `cpu_features` library to v0.7.0. The building option
+  `ENABLE_OWN_CPUFEATURES` has been replaced by `ENABLE_CPUFEATURES`, defaulting
+  to `ON`.
+- Fixed building against GNU Radio v3.10.2.0.
+
+### Improvements in Reliability:
+
+- Fix some defects detected by Coverity Scan 2021.12.1.
+
+### Improvements in Usability:
+
+- Added a script at `src/utils/scripts/download-galileo-almanac.sh` that
+  downloads an XML file with the latest Galileo almanac published by the
+  European GNSS Service Centre at https://www.gsc-europa.eu/product-almanacs
+
+See the definitions of concepts and metrics at
+https://gnss-sdr.org/design-forces/
+
+&nbsp;
+
 ## [GNSS-SDR v0.0.16](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.16) - 2022-02-15
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6090349.svg)](https://doi.org/10.5281/zenodo.6090349)
 
 ### Improvements in Availability:
 
@@ -119,6 +150,8 @@ https://gnss-sdr.org/design-forces/
 &nbsp;
 
 ## [GNSS-SDR v0.0.15](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.15) - 2021-08-23
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5242839.svg)](https://doi.org/10.5281/zenodo.5242839)
 
 ### Improvements in Availability:
 
@@ -244,6 +277,8 @@ https://gnss-sdr.org/design-forces/
 
 ## [GNSS-SDR v0.0.14](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.14) - 2021-01-08
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4428100.svg)](https://doi.org/10.5281/zenodo.4428100)
+
 ### Improvements in Availability:
 
 - Fixed bug in acquisition detection when the configuration parameter
@@ -352,6 +387,8 @@ https://gnss-sdr.org/design-forces/
 &nbsp;
 
 ## [GNSS-SDR v0.0.13](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.13) - 2020-07-29
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3965566.svg)](https://doi.org/10.5281/zenodo.3965566)
 
 ### Improvements in Efficiency:
 
@@ -466,6 +503,8 @@ https://gnss-sdr.org/design-forces/
 &nbsp;
 
 ## [GNSS-SDR v0.0.12](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.12) - 2020-03-13
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3709089.svg)](https://doi.org/10.5281/zenodo.3709089)
 
 ### Improvements in Accuracy:
 
@@ -661,6 +700,8 @@ https://gnss-sdr.org/design-forces/
 
 ## [GNSS-SDR v0.0.11](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.11) - 2019-08-04
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3359989.svg)](https://doi.org/10.5281/zenodo.3359989)
+
 This release has several improvements in different dimensions, addition of new
 features and bug fixes:
 
@@ -814,6 +855,8 @@ https://gnss-sdr.org/design-forces/
 
 ## [GNSS-SDR v0.0.10](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.10) - 2018-12-14
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2279988.svg)](https://doi.org/10.5281/zenodo.2279988)
+
 This release has several improvements in different dimensions, addition of new
 features and bug fixes:
 
@@ -892,8 +935,7 @@ features and bug fixes:
 - Improvements in the RTCM server stability.
 - Improvements in the correctness of generated RINEX files.
 - The receiver can read and make use of Galileo almanac XML files published by
-  the European GNSS Service Centre at
-  https://www.gsc-europa.eu/system-status/almanac-data
+  the European GNSS Service Centre at https://www.gsc-europa.eu/product-almanacs
 - Own-defined XML schemas for navigation data published at
   https://github.com/gnss-sdr/gnss-sdr/tree/next/docs/xml-schemas
 - Added program `rinex2assist` to convert RINEX navigation files into XML files
@@ -1046,7 +1088,7 @@ https://gnss-sdr.org/design-forces/
 
 ## [GNSS-SDR v0.0.9](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.9) - 2017-02-13
 
-DOI: https://doi.org/10.5281/zenodo.291371
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.291371.svg)](https://doi.org/10.5281/zenodo.291371)
 
 This release has several improvements, addition of new features and bug fixes in
 many dimensions:
@@ -1167,7 +1209,7 @@ https://gnss-sdr.org/design-forces/
 
 ## [GNSS-SDR v0.0.8](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.8) - 2016-07-04
 
-DOI: https://doi.org/10.5281/zenodo.57022
+[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.57022.svg)](http://dx.doi.org/10.5281/zenodo.57022)
 
 This is a maintenance and bug fix release with no relevant new features with
 respect to v0.0.7. The main changes are:
@@ -1193,7 +1235,7 @@ respect to v0.0.7. The main changes are:
 
 ## [GNSS-SDR v0.0.7](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.7) - 2016-05-15
 
-DOI: https://doi.org/10.5281/zenodo.51521
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.51521.svg)](https://doi.org/10.5281/zenodo.51521)
 
 This release has several improvements, addition of new features and bug fixes:
 
@@ -1285,6 +1327,8 @@ This release has several improvements, addition of new features and bug fixes:
 
 ## [GNSS-SDR v0.0.6](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.6) - 2015-09-02
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.30104.svg)](https://doi.org/10.5281/zenodo.30104)
+
 This release has several improvements and bug fixes:
 
 - Added initial support to multi-band, multi-source configurations (multiple
@@ -1335,6 +1379,8 @@ This release has several improvements and bug fixes:
 
 ## [GNSS-SDR v0.0.5](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.5) - 2015-01-13
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13920.svg)](https://doi.org/10.5281/zenodo.13920)
+
 This release has several improvements and bug fixes:
 
 - Now GNSS-SDR can be installed on the system with the usual
@@ -1361,6 +1407,8 @@ This release has several improvements and bug fixes:
 &nbsp;
 
 ## [GNSS-SDR v0.0.4](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.4) - 2014-09-08
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.11628.svg)](https://doi.org/10.5281/zenodo.11628)
 
 This release has several improvements and bug fixes:
 
@@ -1398,6 +1446,8 @@ This release has several improvements and bug fixes:
 &nbsp;
 
 ## [GNSS-SDR v0.0.3](https://github.com/gnss-sdr/gnss-sdr/releases/tag/v0.0.3) - 2014-06-30
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10708.svg)](https://doi.org/10.5281/zenodo.10708)
 
 This release has several improvements and bug fixes, completing the transition
 from Subversion to Git. The main changes are: